# Chunked generic slice eq