@the8472 2023-10-04
> Hmm, we used to have a 4-way unroll in all the short-circuiting slice iterator stuff too, because it looks great in simple stuff,

That was just an unroll though, still short-circuiting for each item, right? This is meant to also enable vectorization for a subset of the cases.