@the8472 2023-10-04
Ok, the opt-results have their codegen units perturbed. Bad for comparisons. Maybe we should have some 1CGU benchmarks...

But the check-results are informative. One of the issues is `ValTree` which can be nested slices. So yeah, splitting a short and long case probably makes sense + peeling the first chunk.