@the8472 2023-10-07
This vectorizes nicely on AVX2 (got it down to 12ns/iteration), has a guaranteed preroll that's not chunked in case comparisons are expensive and produces [less unoptimized llvm-ir](https://rust.godbolt.org/z/YczK531bb) than the current implementation. After optimizations it's more IR due to unrolling.

If this won't make perf happy then I'm out of ideas.

@bors try
@rust-timer queue include=html5ever,webrender,bitmaps,helloworld,cargo,hyper,exa,unify-linearly